+++ HISTORY.rst
@@ -17,6 +17,7 @@
   automatically.
 - Use TemplateField for the web link template.
 - Use TemplateField for smart links.
+- Add the ID and the URL to the checkout serializer.
 
 3.3.4 (2019-12-09)
 ==================